refactor(rocm): move the page and activation AITER blocks into _rocm/
Both ops carried their whole AITER path inline in the upstream file, so
every upstream edit near those helpers was a merge conflict for changes
that have nothing to do with the op's shared logic.

Each upstream call site now keeps only a delegation:

    if (result := maybe_silu_and_mul(out, input, backend)) is not None:
        return result

and the helpers, the custom-op/fake-op registration, and the backend
selection move to flashinfer/_rocm/<op>.py. The additive footprint in
the upstream files drops accordingly:

    page.py        +169/-1  ->  +51/-2
    activation.py   +50/-4  ->  +33/-4

page.py evacuates 70%; activation.py only 34%, because its residual is
a genuine backend= signature extension on upstream's own function --
docstring, parameter, and validation -- which cannot move without
forking the function.

Behaviour is unchanged, including the two error paths that are easy to
drop in a move: the explicit backend="aiter" layout/dtype check and the
unknown-backend ValueError both survive in maybe_append_paged_kv_cache.

test_append_paged_kv_cache_aiter_hip.py reached into flashinfer.page for
three moved symbols and had to follow them to flashinfer._rocm.page. The
last of those used a flashinfer._rocm.page.X attribute chain; nothing
imports that submodule eagerly (flashinfer/page.py imports it lazily
inside the call), so it takes an explicit import instead.

Verified on gfx950 (MI355X), ROCm 7.2, aiter 0.1.10: 101 passed, 0
failed, 0 skipped, 0 reruns across the append/activation/dispatch
suites. The pre-fix run of the same three files failed 8.
